+++ CHANGELOG.rst
@@ -1,4 +1,11 @@
 ######
+2.3.0
+######
+
+- Add support for storing the token encrypted in the DB along with an endpoint
+
+
+######
 2.2.0
 ######